Type objects moved to types.ts file. Removed default popups. Changed endpoint method, so it could be scaled later. Removed unused comments in endpoint controller. Modified fetch method moved to utils folder.
