+++ backend/controllers/LocationsController.ts
@@ -1,13 +1,11 @@
 import { Response, Request } from "express";
 import Location from "../models/Locations";
-// import { locations as locConst, locationType } from '../../frontend/src/constants/develop';
-// import {equal} from "assert";
 
 const LocationsController = {
   async getLocationsByZoom(req: Request, res: Response) {
     try {
-      const center: [number, number] = [JSON.parse(req.params.center).lat, JSON.parse(req.params.center).lng],
-          bounds = JSON.parse(req.params.bounds);
+      const center: [number, number] = [JSON.parse(req.body.center).lat, JSON.parse(req.body.center).lng],
+          bounds = JSON.parse(req.body.bounds);
 
       const height = +(bounds._northEast.lat - bounds._southWest.lat);
       const width = +(bounds._northEast.lng - bounds._southWest.lng);
@@ -25,10 +23,6 @@ const LocationsController = {
         _id: l._id,
         coordinates: l.coordinates,
       }));
-
-      // if (!locations || locations.length) {
-      //   return res.status(400).json({ message: "Location doesn't exist" });
-      // }
 
       locations = locations.slice(0, (locations.length < 50)? locations.length : 50);
 
@@ -69,7 +63,7 @@ const LocationsController = {
         const result = await newLocation.save(newLocation as any);
         res.status(200).json(result);
       } else {
-        res.status(500).json({ message: "Data is present" });
+        res.status(400).json({ message: "Data is present" });
       }
     } catch (err: any) {
       return res.status(500).json({ message: err.message });
