+++ frontend/types.ts
@@ -1,0 +1,39 @@
+export type latlngType = {
+  lat: number;
+  lng: number;
+};
+
+export type boundsType = {
+  _northEast: {
+    lat: number;
+    lng: number;
+  };
+  _southWest: {
+    lat: number;
+    lng: number;
+  };
+};
+
+export type locationType = {
+  _id?: string;
+  locationName: string;
+  rating: {
+    likes: number;
+    dislikes: number;
+  };
+  coordinates: [number, number];
+  photoSrc: string;
+  description: string;
+  comments: string[];
+};
+
+export type lightLocationType = {
+  _id?: string;
+  locationName: string;
+  rating: {
+    likes: number;
+    dislikes: number;
+  };
+  coordinates: [number, number];
+  photoSrc: string;
+};
